refactor(cart/CardCart.jsx): fix page.jsx & numberConvent.jsx
