+++ app/cart/page.jsx
@@ -1,5 +1,5 @@
-import products from "./products";
-import Card from './CardCart'
+import products from "./cardCart/products";
+import Card from './cardCart/CardCart'
 function Cart() {
   return (
     <div className="mx-auto w-4/5 md:w-3/5 lg:w-2/5 xl:w-1/3 p-4">
@@ -8,10 +8,14 @@ function Cart() {
         key={product.id}
         id={product.id}
         name={product.name}
+        image={product.image}
         price={product.price}
         stock={product.stock}
         />
       ))}
+      <div>
+        total:
+      </div>
     </div>
   );
 }
